% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_doi_list_altmetrics}
\alias{annotate_doi_list_altmetrics}
\title{Annotate DOI List with altmetrics}
\usage{
annotate_doi_list_altmetrics(doi_list)
}
\arguments{
\item{doi_list}{names of wikipedia category}
}
\value{
dataframee of annotated DOIs with altmetrics
}
\description{
This function get a list of DOIs as input
and create a dataframe of annotated DOIs with altmetrics
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
dois_fetched=unique(unlist(str_match_all(art_test$`*`, doi_regexp)))
annotate_doi_list_altmetrics(list(unique(as.character(dois_fetched))))

}
