% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_doi_list_europmc}
\alias{annotate_doi_list_europmc}
\title{Annotate DOI List with Eurompmc}
\usage{
annotate_doi_list_europmc(doi_list)
}
\arguments{
\item{doi_list}{names of wikipedia category}
}
\value{
dataframe of annotated DOIs with Eurompmc
}
\description{
This function get a list of DOIs as input
and create dataframe of annotated DOIs with Eurompmc
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
dois_fetched=unique(unlist(str_match_all(art_test$`*`, doi_regexp)))
annotate_doi_list_europmc(dois_fetched)
}
