% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_doi_to_bibtex_cross_ref}
\alias{annotate_doi_to_bibtex_cross_ref}
\title{Annotate DOI List with CrossRef}
\usage{
annotate_doi_to_bibtex_cross_ref(doi_list)
}
\arguments{
\item{doi_list}{names of wikipedia category}
}
\value{
bib structure of annotated DOIs with CrossRef
}
\description{
This function get a list of DOIs as input
and create a bib of annotated DOIs with CrossRef
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
dois_fetched=unique(unlist(str_match_all(art_test$`*`, doi_regexp)))
annotate_doi_to_bibtex_cross_ref(dois_fetched)

}
