% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_isbn_google}
\alias{annotate_isbn_google}
\title{Annotate single isbn with google book API}
\usage{
annotate_isbn_google(isbn_nb)
}
\arguments{
\item{isbn_nb}{ISBN number}
}
\value{
dataframe of annotation DOIs from google book API
}
\description{
This function get an ISBN as input
and return a dataframe of annotation DOIs from google book API
}
\examples{


annotate_isbn_google("978-0-15-603135-6")

}
