% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_isbn_list_altmetrics}
\alias{annotate_isbn_list_altmetrics}
\title{Annotate ISBN List with altmetrics}
\usage{
annotate_isbn_list_altmetrics(isbn_list)
}
\arguments{
\item{doi_list}{names of wikipedia category}
}
\value{
dataframee of annotated ISBN with altmetrics
}
\description{
This function get a list of ISBN as input
and create a dataframe of annotated ISBN with altmetrics
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
isbn_fetched=unique(unlist(str_match_all(art_test$`*`, isbn_regexp)))
annotate_isbn_list_altmetrics(list(unique(as.character(isbn_fetched))))

}
