% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{annotate_isbn_openlib}
\alias{annotate_isbn_openlib}
\title{Annotate single ISBN with Open Library}
\usage{
annotate_isbn_openlib(isbn_nb)
}
\arguments{
\item{isbn_nb}{ISBN number}
}
\value{
A data frame with the Open Library response or \code{NULL} if
  the ISBN is not found.
}
\description{
This function retrieves metadata for a given ISBN using the Open
Library API.
}
\examples{
annotate_isbn_openlib("9780156031356")
}
