% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{export_doi_to_bib}
\alias{export_doi_to_bib}
\title{annotate and export DOI list to bibtex file}
\usage{
export_doi_to_bib(doi_list, file_name = "file.bib")
}
\arguments{
\item{doi_list}{list of DOI}

\item{file_name}{output file name}
}
\value{
nothing
}
\description{
This function get a list of DOI as input, annotate it with Rcrossref and export a bibtexfile.
in the working directory.
}
\examples{

category_most_recent=get_category_articles_most_recent(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))
extracted_citation_table=get_regex_citations_in_wiki_table(category_most_recent, "10\\\\.\\\\d{4,9}/[-._;()/:a-z0-9A-Z]+") # doi_regexp
export_doi_to_bib(as.character(extracted_citation_table$citation_fetched)[1:5],"output.bib")

}
