% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{export_extracted_citations_xlsx}
\alias{export_extracted_citations_xlsx}
\title{Export all regex from Article history Table to multiple xlsx with a}
\usage{
export_extracted_citations_xlsx(article_most_recent_table, name_file_prefix)
}
\arguments{
\item{article_most_recent_table}{wiki history or most recent table of multiple wikipedia pages}

\item{name_file_prefix}{output file name prefix}
}
\value{
nothing
}
\description{
This function get a wikipedia article table as input and the name of the target xls file.
an xlsx with the table is written in the working directory.
}
\examples{

category_most_recent=get_category_articles_most_recent(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))
export_extracted_citations_xlsx(category_most_recent, "example")
}
