% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{extract_citations_regexp}
\alias{extract_citations_regexp}
\title{Extract citations using multiple patterns}
\usage{
extract_citations_regexp(article_most_recent_table)
}
\arguments{
\item{article_most_recent_table}{Table of article revisions.}
}
\value{
A named list of data frames of matched citations.
}
\description{
Applies each regular expression stored in \code{pkg.env$regexp_list} to a
table of article revisions.
}
\examples{
art <- get_article_most_recent_table("Zeitgeber")
extract_citations_regexp(art)
}
