% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_anyCount}
\alias{get_anyCount}
\title{Get any regex Count from wikipedia article content}
\usage{
get_anyCount(art_text, regexp)
}
\arguments{
\item{art_text}{text content of wikipedia article}

\item{regexp}{regular expression}
}
\value{
ref_count
}
\description{
This function get a wikipedia article content and a regular expression as arguments
and return regex count in the text.
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
get_anyCount(art_test[9],'<ref.*?</ref>')

}
