% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_article_full_history_table}
\alias{get_article_full_history_table}
\title{Get Article History table}
\usage{
get_article_full_history_table(article_name)
}
\arguments{
\item{article_name}{Path to the input file}
}
\value{
A table with all revisions of the wikipedia article
}
\description{
This function get an english wikipedia article title as input
and retrieve a table containing ids, timestamp, comment,
user, userid,size,content for every revisions of the given wikipedia article
}
\examples{
Zeitgeber_history=get_article_full_history_table("Zeitgeber")
}
