% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_article_initial_table}
\alias{get_article_initial_table}
\title{Get Article Initial Table}
\usage{
get_article_initial_table(article_name)
}
\arguments{
\item{article_name}{Path to the input file}
}
\value{
A table with the initial revision of the wikipedia article
}
\description{
This function get an english wikipedia article title as input
and retrieve a table containing ids, timestamp, comment,
user, userid,size,content of the first revision of the given wikipedia article
}
\examples{
get_article_initial_table("Zeitgeber")
}
