% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_category_articles_creation}
\alias{get_category_articles_creation}
\title{Get Category Articles creation}
\usage{
get_category_articles_creation(list_art)
}
\arguments{
\item{list_art}{list of wikipedia articles}
}
\value{
A table with the creations instance of each wikipedia articles in the input
}
\description{
This function get a list of wikipedia article titles as input
and create a wipipedia creation dates table
}
\examples{

get_category_articles_history=get_category_articles_creation(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))

}
