% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_category_articles_history}
\alias{get_category_articles_history}
\title{Get Category Articles history}
\usage{
get_category_articles_history(list_art)
}
\arguments{
\item{list_art}{list of wikipedia articles}
}
\value{
A table with the all revisions of each wikipedia articles in the input
}
\description{
This function get a list of wikipedia article titles as input
and create a wipipedia history table
}
\examples{

Category_articles_history=get_category_articles_history(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))

}
