% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_category_articles_most_recent}
\alias{get_category_articles_most_recent}
\title{Get Category Articles most recent}
\usage{
get_category_articles_most_recent(list_art)
}
\arguments{
\item{list_art}{list of wikipedia articles}
}
\value{
A table with the most recent instance of each wikipedia articles in the input
}
\description{
This function get a list of wikipedia article titles as input
and create a wipipedia table with most recent version of the pages
}
\examples{

category_most_recent=get_category_articles_most_recent(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))

}
