% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_citation_type}
\alias{get_citation_type}
\title{Get citation type counts}
\usage{
get_citation_type(article_most_recent_table)
}
\arguments{
\item{article_most_recent_table}{Wikipedia revisions table.}
}
\value{
A data frame of citation type counts per article.
}
\description{
This function summarises citation types for each article in a table of
most recent revisions.
}
\examples{
art <- get_article_most_recent_table("Zeitgeber")
get_citation_type(art)
}
