% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_doi_count}
\alias{get_doi_count}
\title{Get DOI Count from wikipedia article content}
\usage{
get_doi_count(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
doi_count
}
\description{
This function get a wikipedia article content
and return doi count
}
\examples{
 text <- "Reference with DOI 10.1000/xyz123"
 get_doi_count(text)

}
