% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_hyperlinkCount}
\alias{get_hyperlinkCount}
\title{Get hyperlinkCount from wikipedia article content}
\usage{
get_hyperlinkCount(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
hyperlinkCount
}
\description{
This function get a wikipedia article content
and return hyperlinkCount
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
get_hyperlinkCount(art_test[9])

}
