% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_page_in_cat_multiple}
\alias{get_page_in_cat_multiple}
\title{Retrieve pages for multiple categories}
\usage{
get_page_in_cat_multiple(catlist, replecement = "_")
}
\arguments{
\item{catlist}{Vector of category names.}

\item{replecement}{Character replacement for spaces.}
}
\value{
Combined data frame of pages.
}
\description{
Retrieve pages for multiple categories
}
\examples{
get_page_in_cat_multiple(c("Category:Biology"))
}
