% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_pagename_in_cat}
\alias{get_pagename_in_cat}
\title{Get Pages Names in Category}
\usage{
get_pagename_in_cat(category)
}
\arguments{
\item{category}{names of wikipedia category}
}
\value{
list of wikipedia pages in the input category
}
\description{
This function get a list of wikipedia article titles as input
and create a wipipedia creation dates table
}
\examples{

get_pagename_in_cat("Circadian rhythm")
# For multiple Categories
unique(unlist(sapply(category_list,get_pagename_in_cat)))

}
