% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_pages_in_cat_table}
\alias{get_pages_in_cat_table}
\title{Retrieve pages in a category}
\usage{
get_pages_in_cat_table(catname, replecement = "_")
}
\arguments{
\item{catname}{Category name.}

\item{replecement}{Character replacement for spaces.}
}
\value{
Data frame of pages in the category.
}
\description{
Returns a table of pages that belong to a specific Wikipedia category.
}
\examples{
get_pages_in_cat_table("Category:Biology")
}
