% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_paresd_citations}
\alias{get_paresd_citations}
\title{Get parsed citations}
\usage{
get_paresd_citations(article_most_recent_table)
}
\arguments{
\item{article_most_recent_table}{wikipedia pages table}
}
\value{
df_cite_parsed_revid_art
}
\description{
This function get a wikipedia page table as input
and return the the parsed citations from the CS1 template with every field.
the output is  6 column dataframe containing the page name, revisionID, type of citation,
an integer value for each citation extracted, the citation variable name (i.e publisher, date, authors)
and the variable value.
}
\examples{
category_most_recent=get_category_articles_most_recent(c("Zeitgeber","Advanced sleep phase disorder","Sleep deprivation"))
paresd_citations=get_paresd_citations(category_most_recent)


}
