% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_pdfs_top20source}
\alias{get_pdfs_top20source}
\title{Plot top citation sources}
\usage{
get_pdfs_top20source(df_cite_parsed_revid_art)
}
\arguments{
\item{df_cite_parsed_revid_art}{Parsed citation data frame.}
}
\value{
No return value, called for side effects.
}
\description{
Generates bar plots of the top 20 values for each citation source type
present in a parsed citation table.
}
\examples{
df <- get_paresd_citations(get_category_articles_most_recent("Zeitgeber"))
get_pdfs_top20source(df)
}
