% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_regex_citations_in_wiki_table}
\alias{get_regex_citations_in_wiki_table}
\title{Get Regex citations in wiki table}
\usage{
get_regex_citations_in_wiki_table(article_wiki_table, citation_regexp)
}
\arguments{
\item{citation_regexp}{citation regular expression}

\item{rticle_wiki_table}{a wiki table}
}
\value{
list of wikipedia pages in the input category
}
\description{
This function get a regex of citations type and a wikipedia article table as input
and th creation dates e table with each matched ciation in each revision/article from a wiki table
}
\details{
accessible regular expression

doi_regexp= "10\\.\\d{4,9}/[-._;()/:a-z0-9A-Z]+"

isbn_regexp='(?<=(isbn|ISBN)\\s?[=:]?\\s?)\\d{1,5}-\\d{1,7}-\\d{1,5}-[\\dX]'

url_regexp = "http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\\(\\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+"

tweet_regexp='\\{\\{cite tweet.*?\\}\\}'

news_regexp='\\{\\{cite news.*?\\}\\}'

journal_regexp='\\{\\{cite journal.*?\\}\\}'

web_regexp='\\{\\{cite web.*?\\}\\}'

article_regexp='\\{\\{cite article.*?\\}\\}'

report_regexp='\\{\\{cite report.*?\\}\\}'

court_regexp='\\{\\{cite court.*?\\}\\}'

press_release_regexp='\\{\\{cite press release.*?\\}\\}'

book_regexp='\\{\\{cite book .*?\\}\\}'

pmid_regexp="(?<=(pmid|PMID)\\s?[=:]\\s?)\\d{5,9}"

ref_regexp='<ref>\\{\\{[c|C]ite.*?\\}\\}</ref>' # in-text refs!

cite_regexp='\\{\\{[c|C]ite.*?\\}\\}'
}
\examples{
Zeitgeber_history=get_article_full_history_table("Zeitgeber")
citations_in_wiki_table=get_regex_citations_in_wiki_table(Zeitgeber_history,doi_regexp)


}
