% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{Get_sci_score}
\alias{Get_sci_score}
\title{Get SciScore from wikipedia article content}
\usage{
Get_sci_score(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
SciScore
}
\description{
This function get a wikipedia article content
and return SciScore
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
Get_sci_score(art_test[9])

}
