% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{Get_sci_score2}
\alias{Get_sci_score2}
\title{Get SciScore2 from wikipedia article content}
\usage{
Get_sci_score2(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
SciScore2
}
\description{
This function get a wikipedia article content
and return SciScore2 doi over refs
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
Get_sci_score2(art_test[9])

}
