% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{Get_source_type_counts}
\alias{Get_source_type_counts}
\title{Get Source Type count}
\usage{
Get_source_type_counts(art_text)
}
\arguments{
\item{art_text}{wiki article content}
}
\value{
citation_type_count
}
\description{
This function get a wikipedia article content as input
and return the the count of a citation from the CS1 template such as book, website, newspaper, journal
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
Get_source_type_counts(art_test[9])


}
