% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_subcat_multiple}
\alias{get_subcat_multiple}
\title{Retrieve subcategories for multiple categories}
\usage{
get_subcat_multiple(catlist, replecement = "_")
}
\arguments{
\item{catlist}{Vector of category names.}

\item{replecement}{Character replacement for spaces.}
}
\value{
Combined data frame of subcategories.
}
\description{
Retrieve subcategories for multiple categories
}
\examples{
get_subcat_multiple(c("Category:Biology"))
}
