% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_subcat_table}
\alias{get_subcat_table}
\title{Retrieve subcategories}
\usage{
get_subcat_table(catname, replecement = "_")
}
\arguments{
\item{catname}{Category name.}

\item{replecement}{Character replacement for spaces.}
}
\value{
Data frame of subcategories.
}
\description{
Returns a table of subcategories for a given Wikipedia category.
}
\examples{
get_subcat_table("Category:Biology")
}
