% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_subcat_with_depth}
\alias{get_subcat_with_depth}
\title{Retrieve subcategories with depth}
\usage{
get_subcat_with_depth(catname, depth, replecement = "_")
}
\arguments{
\item{catname}{Category name.}

\item{depth}{Depth to traverse.}

\item{replecement}{Character replacement for spaces.}
}
\value{
Data frame of subcategories up to the desired depth.
}
\description{
Recursively fetches subcategories up to a specified depth.
}
\examples{
get_subcat_with_depth("Category:Biology", 1)
}
