% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_top_cited_wiki_papers}
\alias{get_top_cited_wiki_papers}
\title{Get top cited Wikipedia papers}
\usage{
get_top_cited_wiki_papers(df_doi_revid_art)
}
\arguments{
\item{df_doi_revid_art}{Data frame linking DOIs to Wikipedia articles.}
}
\value{
A data frame of the top cited DOIs with annotations.
}
\description{
Identifies the most frequently cited DOIs and annotates them with
bibliographic information.
}
\examples{
df <- get_regex_citations_in_wiki_table(get_article_most_recent_table("Zeitgeber"),
                                        "10\\\\.\\\\d{4,9}/[-._;()/:a-z0-9A-Z]+")
get_top_cited_wiki_papers(df)
}
