% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{page_edit_plot}
\alias{page_edit_plot}
\title{Plot page edits}
\usage{
page_edit_plot(
  article_name,
  ymax = NA,
  start = "2020010100",
  end = "2020050100"
)
}
\arguments{
\item{article_name}{Wikipedia article title.}

\item{ymax}{Optional y-axis maximum.}

\item{start}{Start date in YYYYMMDDHH format.}

\item{end}{End date in YYYYMMDDHH format.}
}
\value{
No return value, generates a plot.
}
\description{
Displays the number of edits per week for a given article.
}
\examples{
page_edit_plot("Zeitgeber")
}
