% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{page_view_plot}
\alias{page_view_plot}
\title{Plot page views}
\usage{
page_view_plot(
  article_name,
  ymax = NA,
  start = "2020010100",
  end = "2020050100"
)
}
\arguments{
\item{article_name}{Wikipedia article title.}

\item{ymax}{Optional y-axis maximum.}

\item{start}{Start date in YYYYMMDDHH format.}

\item{end}{End date in YYYYMMDDHH format.}
}
\value{
No return value, generates a plot.
}
\description{
Generates an area plot of Wikipedia page views over time for a given
article.
}
\examples{
page_view_plot("Zeitgeber")
}
