% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{parse_article_ALL_citations}
\alias{parse_article_ALL_citations}
\title{Parse wikipedia article content for ALL citations}
\usage{
parse_article_ALL_citations(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
citations
}
\description{
This function get a wikipedia article content
and return every citations including, news, web ,journal, etc.
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
citation_extracted=parse_article_ALL_citations(art_test[9]))


}
