% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{parse_cite_type}
\alias{parse_cite_type}
\title{Parse Citation Type}
\usage{
parse_cite_type(citation)
}
\arguments{
\item{citation}{extracted citaion from wiki article content}
}
\value{
citation type
}
\description{
This function get a citation as input
and return the type of a citation such as book, website,newspaper, journal
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
citation_type_extracted=as.character(sapply(extract_citations(art_test[9]),replace_wikihypelinks))


}
