% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{plot_article_creation_per_year}
\alias{plot_article_creation_per_year}
\title{Plot article creation over time}
\usage{
plot_article_creation_per_year(article_initial_table, name_title, Cumsum = T)
}
\arguments{
\item{article_initial_table}{Table of initial revisions for articles.}

\item{name_title}{Plot title.}

\item{Cumsum}{Logical indicating whether to plot cumulative counts.}
}
\value{
No return value, generates a plot.
}
\description{
Creates a cumulative or per-year plot of article creation dates.
}
\examples{
initial <- get_article_initial_table("Zeitgeber")
plot_article_creation_per_year(initial, "Zeitgeber")
}
