% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{plot_distribution_source_type}
\alias{plot_distribution_source_type}
\title{Plot distribution of citation source types}
\usage{
plot_distribution_source_type(df_cite_count_revid_art)
}
\arguments{
\item{df_cite_count_revid_art}{Data frame of citation type counts.}
}
\value{
No return value, generates a plot.
}
\description{
Creates a boxplot showing counts of selected citation types.
}
\examples{
df <- get_citation_type(get_article_most_recent_table("Zeitgeber"))
plot_distribution_source_type(df)
}
