% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{plot_navi_timeline}
\alias{plot_navi_timeline}
\title{Plot interactive article timeline}
\usage{
plot_navi_timeline(article_initial_table_sel, article_info_table)
}
\arguments{
\item{article_initial_table_sel}{Subset of initial article table.}

\item{article_info_table}{Table with article page information.}
}
\value{
No return value, launches a Shiny app.
}
\description{
Creates an interactive timevis timeline linking to articles.
}
\examples{
initial <- get_article_initial_table("Zeitgeber")
info <- get_article_info_table("Zeitgeber")
plot_navi_timeline(initial, info)
}
