% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{plot_static_timeline}
\alias{plot_static_timeline}
\title{Plot static article timeline}
\usage{
plot_static_timeline(article_initial_table_sel)
}
\arguments{
\item{article_initial_table_sel}{Subset of initial article table.}
}
\value{
No return value, generates a plot.
}
\description{
Displays article creation dates in a static ggplot timeline.
}
\examples{
initial <- get_article_initial_table("Zeitgeber")
plot_static_timeline(initial)
}
