% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{plot_top_source}
\alias{plot_top_source}
\title{Plot top values for a citation variable}
\usage{
plot_top_source(df_cite_parsed_revid_art, source_type)
}
\arguments{
\item{df_cite_parsed_revid_art}{Parsed citation data frame.}

\item{source_type}{Citation variable to summarise.}
}
\value{
No return value, generates a plot.
}
\description{
Displays the top 20 occurrences of a given citation variable.
}
\examples{
df <- get_paresd_citations(get_category_articles_most_recent("Zeitgeber"))
plot_top_source(df, "publisher")
}
