% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{replace_wikihypelinks}
\alias{replace_wikihypelinks}
\title{Replace wikihypelinks}
\usage{
replace_wikihypelinks(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
art_text
}
\description{
This function get a wikipedia article content
and replace wikihypelinks to clean wikipedia article content
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
citation_type_extracted=as.character(sapply(extract_citations(art_test[9]),replace_wikihypelinks))


}
