% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiHistoRyfunctions.r
\name{get_urlCount}
\alias{get_urlCount}
\title{Get urlCount from wikipedia article content}
\usage{
get_urlCount(art_text)
}
\arguments{
\item{art_text}{text content of wikipedia article}
}
\value{
urlCount
}
\description{
This function get a wikipedia article content
and return urlCount
}
\examples{
art_test=get_article_most_recent_table("Zeitgeber")
get_urlCount(art_test[9])

}
